+++ src/tquality_core/__init__.py
@@ -10,7 +10,9 @@ from tquality_core.services.logger import (
     set_logger_resolver,
     step,
 )
+from tquality_core.utils.os_utils import OSUtils
 from tquality_core.utils.string_utils import StringUtils
+from tquality_core.utils.xpath_utils import XPathUtils
 
 __all__ = [
     "BaseConfig",
@@ -19,9 +21,11 @@ __all__ = [
     "Locator",
     "Logger",
     "LogLevel",
+    "OSUtils",
     "ScreencastProvider",
     "ScreenshotProvider",
     "StringUtils",
+    "XPathUtils",
     "set_logger_resolver",
     "step",
 ]