+++ src/tquality_core/utils/__init__.py
@@ -1,3 +1,5 @@
+from tquality_core.utils.os_utils import OSUtils
 from tquality_core.utils.string_utils import StringUtils
+from tquality_core.utils.xpath_utils import XPathUtils
 
-__all__ = ["StringUtils"]
+__all__ = ["OSUtils", "StringUtils", "XPathUtils"]